Peer review fixes
