+++ mqtt-haskell.cabal
@@ -17,7 +17,7 @@ extra-source-files:
 library
     hs-source-dirs: lib
     exposed-modules: Utils, Broker, Client, Packets
-    other-modules: Packets.Simple, Packets.Builder, Packets.Parser, Packets.Abstract
+    other-modules: Packets.Simple, Packets.Builder, Packets.Parser, Packets.Abstract, Utils.Bits, Utils.BitParsing
     build-depends: base ^>=4.16.4.0, network >= 0.1.0.1, bytestring, containers >= 0.6.5.1
     default-language: Haskell2010
